modified:   workspaces/DummyRosterRC0.1/DummyRoster.Common.EntityModel/Models/Product.cs

+++ workspaces/DummyRosterRC0.1/DummyRoster.Common.EntityModel/Models/Product.cs
@@ -13,7 +13,7 @@ namespace DummyRoster.Common.EntityModel.Models;
 public partial class Product
 {
   [Key]
-  public long Id { get; set; }
+  public int Id { get; set; }
 
   [Column(TypeName = "varchar(32)")]
   public string Name { get; set; } = null!;
@@ -23,10 +23,10 @@ public partial class Product
   public byte[]? Picture { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? CategoryId { get; set; }
+  public int? CategoryId { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? SupplierId { get; set; }
+  public int? SupplierId { get; set; }
 
   [Column(TypeName = "varchar(24)")]
   public string? QuantityPerUnit { get; set; }
@@ -35,13 +35,13 @@ public partial class Product
   public byte[]? UnitPrice { get; set; }
 
   [Column(TypeName = "SMALLINT")]
-  public long? UnitsInStock { get; set; }
+  public int? UnitsInStock { get; set; }
 
   [Column(TypeName = "SMALLINT")]
-  public long? UnitsOnOrder { get; set; }
+  public int? UnitsOnOrder { get; set; }
 
   [Column(TypeName = "SMALLINT")]
-  public long? ReorderLevel { get; set; }
+  public int? ReorderLevel { get; set; }
 
   [Column(TypeName = "BOOLEAN")]
   public byte[] Discontinued { get; set; } = null!;
